+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Cruxite dowel emerging from a cruxtruder now breaks if the pipe below breaks
 - The cruxtruder now drops any held raw cruxite when destroyed
 - The totem lathe card slot block will now drop held cards when destroyed
+- The cruxite dowel block in the totem lathe now drops the dowel when destroyed
 - The totem lathe rod now also updates if the connected dowel is destroyed
 
 ### Fixed